+++ cpp/pybind/core/core.cpp
@@ -35,107 +35,6 @@
 namespace open3d {
 namespace core {
 
-/// Convert Tensor class to py::array (Numpy array).
-py::array TensorToPyArray(const Tensor& tensor) {
-    if (tensor.GetDevice().GetType() != Device::DeviceType::CPU) {
-        utility::LogError(
-                "Can only convert CPU Tensor to numpy. Copy Tensor to CPU "
-                "before converting to numpy.");
-    }
-    py::dtype py_dtype =
-            py::dtype(pybind_utils::DtypeToArrayFormat(tensor.GetDtype()));
-    py::array::ShapeContainer py_shape(tensor.GetShape());
-    SizeVector strides = tensor.GetStrides();
-    int64_t element_byte_size = tensor.GetDtype().ByteSize();
-    for (auto& s : strides) {
-        s *= element_byte_size;
-    }
-    py::array::StridesContainer py_strides(strides);
-
-    // `base_tensor` is a shallow copy of `tensor`. `base_tensor`
-    // is on the heap and is owned by py::capsule
-    // `base_tensor_capsule`. The capsule is referenced as the
-    // "base" of the numpy tensor returned by o3d.Tensor.numpy().
-    // When the "base" goes out-of-scope (e.g. when all numpy
-    // tensors referencing the base have gone out-of-scope), the
-    // deleter is called to free the `base_tensor`.
-    //
-    // This behavior is important when the origianl `tensor` goes
-    // out-of-scope while we still want to keep the data alive.
-    // e.g.
-    //
-    // ```python
-    // def get_np_tensor():
-    //     o3d_t = o3d.Tensor(...)
-    //     return o3d_t.numpy()
-    //
-    // # Now, `o3d_t` is out-of-scope, but `np_t` still
-    // # references the base tensor which references the
-    // # underlying data of `o3d_t`. Thus np_t is still valid.
-    // # When np_t goes out-of-scope, the underlying data will be
-    // # finally freed.
-    // np_t = get_np_tensor()
-    // ```
-    //
-    // See:
-    // https://stackoverflow.com/questions/44659924/returning-numpy-arrays-via-pybind11
-    Tensor* base_tensor = new Tensor(tensor);
-
-    // See PyTorch's torch/csrc/Module.cpp
-    auto capsule_destructor = [](PyObject* data) {
-        Tensor* base_tensor = reinterpret_cast<Tensor*>(
-                PyCapsule_GetPointer(data, "open3d::Tensor"));
-        if (base_tensor) {
-            delete base_tensor;
-        } else {
-            PyErr_Clear();
-        }
-    };
-
-    py::capsule base_tensor_capsule(base_tensor, "open3d::Tensor",
-                                    capsule_destructor);
-    return py::array(py_dtype, py_shape, py_strides, tensor.GetDataPtr(),
-                     base_tensor_capsule);
-}
-
-Tensor PyArrayToTensor(py::array array, bool inplace) {
-    py::buffer_info info = array.request();
-
-    SizeVector shape(info.shape.begin(), info.shape.end());
-    SizeVector strides(info.strides.begin(), info.strides.end());
-    for (size_t i = 0; i < strides.size(); ++i) {
-        strides[i] /= info.itemsize;
-    }
-    Dtype dtype = pybind_utils::ArrayFormatToDtype(info.format, info.itemsize);
-    Device device("CPU:0");
-
-    array.inc_ref();
-    std::function<void(void*)> deleter = [array](void*) -> void {
-        py::gil_scoped_acquire acquire;
-        array.dec_ref();
-    };
-    auto blob = std::make_shared<Blob>(device, info.ptr, deleter);
-    Tensor t_inplace(shape, strides, info.ptr, dtype, blob);
-
-    if (inplace) {
-        return t_inplace;
-    } else {
-        return t_inplace.Copy();
-    }
-}
-
-Tensor PyListToTensor(const py::list& list) {
-    py::object numpy = py::module::import("numpy");
-    py::array np_array = numpy.attr("array")(list);
-    return PyArrayToTensor(np_array, false);
-}
-
-Tensor PyTupleToTensor(const py::tuple& tuple) {
-    py::object numpy = py::module::import("numpy");
-    py::array np_array = numpy.attr("array")(tuple);
-    return PyArrayToTensor(np_array, false);
-}
-
 void pybind_core(py::module& m) {
     py::module m_core = m.def_submodule("core");
