+++ cpp/pybind/core/core.h
@@ -32,55 +32,6 @@
 namespace open3d {
 namespace core {
 
-/// Converts Tensor to py::array (Numpy array). The python object holds a
-/// reference to the Tensor, and when it goes out of scope, the Tensor's
-/// reference counter will be decremented by 1.
-///
-/// You may use this helper function for exporting data to Numpy.
-///
-/// To expose a C++ buffer to python, we need to carefully manage the buffer
-/// ownership. You firest need to allocate the memory in the heap (e.g. with
-/// `new`, `malloc`, avoid using containers that frees up memory when the C++
-/// variable goes out of scope), then in pybind11, define a deleter function for
-/// py::array_t that deallocates the buffer. This deleater function will be
-/// called once the python reference count decreases to 0. See
-/// https://stackoverflow.com/a/44682603/1255535 for details. This approach is
-/// efficient since no memory copy is required.
-///
-/// Alternatively, you can create a Tensor with a **copy** of your data (so that
-/// your original buffer can be freed), and let TensorToPyArray generate a
-/// py::array that manages the buffer lifetime automatically. This is more
-/// convienent, but will require an extra copy.
-py::array TensorToPyArray(const Tensor& tensor);
-
-/// Converts py::array (Numpy array) to Tensor.
-///
-/// You may use this helper function for importing data from Numpy.
-///
-/// \param inplace If True, Tensor will directly use the underlying Numpy
-/// buffer. However, The data will become invalid once the Numpy variable is
-/// deallocated, the Tensor's data becomes invalid without notice. If False, the
-/// python buffer will be copied.
-Tensor PyArrayToTensor(py::array array, bool inplace);
-
-/// Converts py::list to Tensor.
-///
-/// Nested lists are supported, e.g. [[0, 1, 2], [3, 4, 5]] becomes a 2x3
-/// tensor. For "ragged" list of invalid shapes, e.g. ((0, 1, 2, 3, 4, 5), (2,
-/// 3)), the np_array's dtype is "O", a proper exception will be thrown.
-///
-/// The dtype will be inferred from the value of the list.
-Tensor PyListToTensor(const py::list& list);
-
-/// Converts py::tuple to Tensor.
-///
-/// Nested tuples are supported, e.g. ((0, 1, 2), (3, 4, 5)) becomes a 2x3
-/// tensor. For "ragged" tuple of invalid shapes, e.g. ((0, 1, 2, 3, 4, 5), (2,
-/// 3)), the np_array's dtype is "O", a proper exception will be thrown.
-///
-/// The dtype will be inferred from the value of the tuple.
-Tensor PyTupleToTensor(const py::tuple& tuple);
-
 void pybind_core(py::module& m);
 void pybind_cuda_utils(py::module& m);
 void pybind_core_blob(py::module& m);
