+++ cpp/pybind/core/tensor_accessor.cpp
@@ -30,6 +30,7 @@
 #include "open3d/core/TensorKey.h"
 #include "open3d/utility/Optional.h"
 #include "pybind/core/core.h"
+#include "pybind/core/tensor_converter.h"
 #include "pybind/docstring.h"
 #include "pybind/open3d_pybind.h"
 #include "pybind/pybind_utils.h"
