+++ cpp/pybind/ml/contrib/contrib_subsample.cpp
@@ -25,7 +25,7 @@
 // ----------------------------------------------------------------------------
 
 #include "open3d/ml/contrib/GridSubsampling.h"
-#include "pybind/core/core.h"
+#include "pybind/core/tensor_converter.h"
 #include "pybind/docstring.h"
 #include "pybind/ml/contrib/contrib.h"
 #include "pybind/open3d_pybind.h"
